Seperated the utlity methods in utils.py
